// src/server/enums/status.ts
export enum MemberStatus {
    Paid = 1,
    Lifetime = 2,
    BulletinOnly = 3,
    Deceased = 4,
    DroppedOut = 5,
    Inactive = 6,
    RenewalPending = 7,
    TermByBoard = 8,
    WaitingList = 9,
    Widow = 10,
    Undetermined = 11,
    Banned = 12,
    NonPayment = 13,
    Vendor = 14,
    Guest = 15,
    MembershipDeclined = 16,
    OnHold = 17,
    ApplicationSubmitted = 18,
    Inducted = 19,
    ApplicationRequested = 20,
    ReinstatementGranted = 21,
    NotInGoodStanding = 22
}